pdf-x-agent/requirements.txt
# Core dependencies
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
tweepy==4.14.0
chromadb==0.4.22
sentence-transformers==2.2.2
ollama==0.3.3
numpy==1.24.3
pandas==2.0.3
beautifulsoup4==4.12.2
lxml==4.9.3
transformers==4.36.2
torch==2.1.0
schedule==1.2.0
pyyaml==6.0.1
requests==2.31.0